requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyJWT==2.8.0
cryptography==41.0.7
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
django-extensions==3.2.3
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Redis for caching and sessions
django-redis==5.4.0
redis==5.0.1

# Security enhancements
django-ratelimit==4.1.0
django-password-strength==1.2.1

# Google Cloud Secret Manager
google-cloud-secret-manager==2.18.1

# Code quality tools (development)
flake8==6.1.0
black==23.12.1
isort==5.13.2

# Testing and coverage
coverage==7.3.4

# Logging
python-json-logger==2.0.7

# Monitoring and metrics
prometheus-client==0.19.0

# API Documentation
drf-spectacular==0.27.0
